+++ docs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @stackframe/docs
 
+## 2.7.6
+
+### Patch Changes
+
+- Fixed bugs, updated Neon requirements
+
+## 2.7.5
+
+### Patch Changes
+
+- Various changes
+
 ## 2.7.4
 
 ## 2.7.3