+++ examples/cjs-test/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @stackframe/example-cjs-test
 
+## 2.7.6
+
+### Patch Changes
+
+- Fixed bugs, updated Neon requirements
+- Updated dependencies
+  - @stackframe/stack@2.7.6
+
+## 2.7.5
+
+### Patch Changes
+
+- @stackframe/stack@2.7.5
+
 ## 2.7.4
 
 ### Patch Changes