+++ examples/docs-examples/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @stackframe/docs-examples
 
+## 2.7.6
+
+### Patch Changes
+
+- Fixed bugs, updated Neon requirements
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.6
+  - @stackframe/stack-ui@2.7.6
+  - @stackframe/stack@2.7.6
+
+## 2.7.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.5
+  - @stackframe/stack@2.7.5
+  - @stackframe/stack-ui@2.7.5
+
 ## 2.7.4
 
 ### Patch Changes