+++ packages/init-stack/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stackframe/init-stack
 
+## 2.7.6
+
+### Patch Changes
+
+- Fixed bugs, updated Neon requirements
+
+## 2.7.5
+
 ## 2.7.4
 
 ## 2.7.3